requirements.txt
setuptools
pandas
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
lightgbm
tqdm
pyyaml
-e .
